requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.7.3
pytest==8.2.1
pytest-asyncio==0.24.0
httpx==0.27.0
python-multipart==0.0.9
Pillow==10.3.0
stability-sdk==0.4.0
redis==5.2.0
prometheus-client==0.21.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
